Source formatting
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@33411 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/security/auth/OpenIdAutoLogin.java
@@ -64,8 +64,7 @@ public class OpenIdAutoLogin implements AutoLogin {
 
 			session.removeAttribute(WebKeys.OPEN_ID_LOGIN);
 
-			User user = UserLocalServiceUtil.getUserById(
-				userId.longValue());
+			User user = UserLocalServiceUtil.getUserById(userId);
 
 			credentials = new String[3];
